# openapi.yaml
openapi: 3.1.0
info:
  title: Sans Limites — API
  description: >
    API du réseau social sportif Sans Limites.
    L'authentification se fait par JWT : récupérez un `accessToken` via
    /auth/register ou /auth/login, puis envoyez-le dans l'en-tête
    `Authorization: Bearer <token>` sur les routes protégées.
  version: 0.1.0
servers:
  - url: http://localhost:8080/api
    description: Cluster k3d (Ingress Traefik)
  - url: http://localhost:3001/api
    description: Exécution locale (npm run start:dev)
...
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      tags: [auth]
      summary: Inscription
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterInput' }
      responses:
        '201':
          description: Compte créé, tokens émis
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '409': { description: Email ou nom d'utilisateur déjà utilisé }
  /auth/login:
    post:
      tags: [auth]
      summary: Connexion
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginInput' }
      responses:
        '200':
          description: Connecté, tokens émis
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { description: Identifiants invalides }
...
  /auth/refresh:
    post:
      tags: [auth]
      summary: Renouvellement des tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, refreshToken]
              properties:
                userId: { type: string }
                refreshToken: { type: string }
      responses:
        '200':
          description: Nouveaux tokens
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { description: Refresh token invalide }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterInput:
      type: object
      required: [email, password, username, name]
      properties:
        email: { type: string, format: email, example: alex@sans-limite.dev }
        password: { type: string, minLength: 8, example: motdepasse123 }
        username: { type: string, example: alex.martin }
        name: { type: string, example: Alexandre Martin }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    TokenPair:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
...